requirements.txt
Flask>=2.0
flask-cors
pandas
openpyxl
python-dotenv